♻️ Components seperation
